cloud_functions/load-v1-tournament-data-1/requirements.txt
numerapi==2.8.1
pandas==1.3.2
google-cloud-storage==1.18.1
google-cloud-bigquery==1.21.0
google-cloud-pubsub==2.8.0
pyarrow==3.0.0
